added support for cookies and session

+++ routes/UserRoute.js
@@ -8,7 +8,8 @@ module.exports = (app) => {
         UserService.getByUserTz(userDetails)
             .then(user => {
                 delete user.psw;
-                // req.session.loggedinUser = user;
+                req.session.cookie.maxAge = 60 * 60 * 1000;
+                req.session.loggedinUser = user;
                 res.json(user);
             })
             .catch(err => res.status(401).send('Wrong User/Password'));
@@ -16,11 +17,14 @@ module.exports = (app) => {
 
     //Query
     app.get(USER_URL, (req, res) => {
+        if (!req.session.loggedinUser) return;
         UserService.query()
             .then(users => res.json(users))
     })
     //get By Class
     app.get(USER_URL + '/byClass', (req, res) => {
+        const { type } = req.session.loggedinUser;
+        if (!req.session.loggedinUser || (type !== 'h' && type !== 't')) return;
         const classCode = +req.query.classCode;
         const schoolCode = +req.query.schoolCode;
         UserService.getByClassCode(classCode, schoolCode)
@@ -30,6 +34,7 @@ module.exports = (app) => {
 
     //getById
     app.get(USER_URL + '/:userId', (req, res) => {
+        if (!req.session.loggedinUser) return;
         const userId = req.params.userId
         UserService.getById(userId)
             .then(user => {
@@ -39,6 +44,7 @@ module.exports = (app) => {
 
     //updateUser
     app.put(USER_URL + '/:userId', (req, res) => {
+        if (!req.session.loggedinUser) return;
         let updatedUser = req.body;
         UserService.update(updatedUser)
             .then(user => {
@@ -48,6 +54,7 @@ module.exports = (app) => {
 
     //unshiftSubmission
     app.put(`${USER_URL}/submission/:userId`, (req, res) => {
+        if (!req.session.loggedinUser) return;
         const userId = req.params.userId;
         const submission = req.body;
         UserService.addSubmission(userId, submission)
@@ -57,6 +64,7 @@ module.exports = (app) => {
 
     //delete
     app.delete(USER_URL + ':/userId', (req, res) => {
+        if (!req.session.loggedinUser) return;
         const userId = req.params.userId;
         UserService.remove(userId)
             .then(() => res.end(`User ${userId} was Deleted successfully`));
@@ -64,6 +72,7 @@ module.exports = (app) => {
 
     //add
     app.post(USER_URL, (req, res) => {
+        if (!req.session.loggedinUser) return;
         const user = req.body;
         UserService.add(user)
             .then(user => {
